py_backend/requirements.txt
# Vercel Serverless Python Dependencies
Flask==3.0.2
Werkzeug==3.0.1
supabase==2.3.4
pandas==2.2.0
scipy==1.12.0
numpy==1.26.4
requests==2.31.0
beautifulsoup4==4.12.3
python-dotenv==1.0.1
